requirements.txt
flask
ezdxf
pymupdf
pytesseract
pillow
pandas
openpyxl